Removed FundPolicy. Corrected fund update rule.
We are not worried about authorization yet, so we are removing the FundPolicy class, which didn't seem to be working correctly anyway.  Correct the rule for updating funds so that the 'unique' rule on the 'symbol' field doesn't prevent us from updating that fund itself, merely prevents duplicating a symbol.